reduce size of large file
