+++ enclone_main/src/lib.rs
@@ -11,3 +11,4 @@ pub mod merge_onesies;
 pub mod populate_features;
 pub mod setup;
 pub mod split_orbits;
+pub mod vars;
