+++ enclone_main/src/main_enclone.rs
@@ -13,6 +13,7 @@ use crate::merge_onesies::*;
 use crate::populate_features::*;
 use crate::setup::*;
 use crate::split_orbits::*;
+use crate::vars::*;
 use debruijn::dna_string::DnaString;
 use enclone::allele::*;
 use enclone::explore::*;
@@ -36,7 +37,6 @@ use io_utils::*;
 use perf_stats::*;
 use pretty_trace::*;
 use rayon::prelude::*;
-use regex::Regex;
 use serde_json::Value;
 use stats_utils::*;
 use std::{
@@ -220,111 +220,7 @@ pub fn main_enclone(args: &Vec<String>) {
 
     // Find matching features for <regular expression>_g etc.
 
-    let tstar = Instant::now();
-    ctl.clono_print_opt.regex_match =
-        vec![HashMap::<String, Vec<usize>>::new(); ctl.origin_info.n()];
-    let ends0 = [
-        "_g", "_ab", "_cr", "_cu", "_g_μ", "_ab_μ", "_cr_μ", "_cu_μ", "_g_%",
-    ];
-    let ends1 = ["_g", "_ab", "_cr", "_cu", "_g", "_ab", "_cr", "_cu", "_g"];
-    let suffixes = ["", "_min", "_max", "_μ", "_Σ"];
-    let mut ends = Vec::<String>::new();
-    let mut endsz = Vec::<String>::new();
-    for (ix, x) in ends0.iter().enumerate() {
-        for y in suffixes.iter() {
-            ends.push(format!("{}{}", x, y));
-            endsz.push(ends1[ix].to_string());
-        }
-    }
-    let mut vars = ctl.clono_print_opt.lvars.clone();
-    vars.append(&mut ctl.parseable_opt.pcols.clone());
-    unique_sort(&mut vars);
-    ctl.perf_stats(&tstar, "doing miscellaneous stuff");
-    let tomega = Instant::now();
-    for x in vars.iter() {
-        for (iy, y) in ends.iter().enumerate() {
-            let mut xc = x.clone();
-            if x.ends_with("_cell") {
-                xc = xc.rev_before("_cell").to_string();
-            }
-            if xc.ends_with(y) {
-                let mut p = xc.rev_before(y);
-                if p.contains(':') {
-                    p = p.after(":");
-                }
-                let pp = format!("{}{}", p, endsz[iy]);
-                if !p.is_empty() && Regex::new(&p).is_ok() {
-                    let mut ok = true;
-                    let mut px = false;
-                    let b = p.as_bytes();
-                    for i in 0..p.len() {
-                        if !((b[i] >= b'A' && b[i] <= b'Z')
-                            || (b[i] >= b'a' && b[i] <= b'z')
-                            || (b[i] >= b'0' && b[i] <= b'9')
-                            || b".-_[]()|*".contains(&b[i]))
-                        {
-                            ok = false;
-                            break;
-                        }
-                        if b"[]()|*".contains(&b[i]) {
-                            px = true;
-                        }
-                    }
-                    if ok && px {
-                        let reg = Regex::new(&format!("^{}$", p));
-                        for li in 0..ctl.origin_info.n() {
-                            let mut js = Vec::<usize>::new();
-                            for j in 0..gex_info.gex_features[li].len() {
-                                let f = &gex_info.gex_features[li][j];
-                                let ff = f.split('\t').collect::<Vec<&str>>();
-                                let mut ok = false;
-                                if ff[2].starts_with("Antibody") {
-                                    if y.contains("_ab") {
-                                        ok = true;
-                                    }
-                                } else if ff[2].starts_with("CRISPR") {
-                                    if y.contains("_cr") {
-                                        ok = true;
-                                    }
-                                } else if ff[2].starts_with("Custom") {
-                                    if y.contains("_cu") {
-                                        ok = true;
-                                    }
-                                } else if y.contains("_g") {
-                                    ok = true;
-                                }
-                                if ok
-                                    && (reg.as_ref().unwrap().is_match(&ff[0])
-                                        || reg.as_ref().unwrap().is_match(&ff[1]))
-                                {
-                                    js.push(j);
-                                }
-                            }
-                            if js.len() > 0 {
-                                ctl.clono_print_opt.regex_match[li].insert(pp.clone(), js);
-                            }
-                        }
-                        let mut matches = false;
-                        for li in 0..ctl.origin_info.n() {
-                            if ctl.clono_print_opt.regex_match[li].contains_key(&pp) {
-                                matches = true;
-                            }
-                        }
-                        if !matches {
-                            eprintln!(
-                                "\nLead variable {} contains a pattern that matches \
-                                no features.\n",
-                                x
-                            );
-                            std::process::exit(1);
-                        }
-                        break;
-                    }
-                }
-            }
-        }
-    }
-    ctl.perf_stats(&tomega, "messing with variables");
+    match_vars(&mut ctl, &gex_info);
 
     // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
